python_backnd/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# MongoDB
motor==3.3.2
pymongo==4.6.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP Client (for OAuth)
httpx==0.26.0

# Email (optional - for email OTP)
# aiosmtplib==3.0.1

# SMS (optional - for phone OTP via Twilio)
# twilio==8.10.0

# Environment
python-dotenv==1.0.0
